add missing imports
